Move the versioned tracked into its own file
